Add delete create and update to trip

+++ trips/src/main/java/trips/web/controller/TripController.java
@@ -14,6 +14,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -41,20 +42,11 @@ public class TripController {
 	@Autowired
 	private RouteDao routeDao;
 
-	@RequestMapping("/trip/list.json")
-	public void list(HttpServletResponse response, HttpServletRequest request)
+	@RequestMapping("/trip/list/{userId}")
+	public void list(@PathVariable Integer userId, HttpServletResponse response, HttpServletRequest request)
 			throws JSONException, IOException {
-		JSONObject reqJSON = null;
 		List<Trip> trips = null;
-		
-		int userId = 0;
-		try {
-			reqJSON = new JSONObject(URLDecoder.decode(request.getQueryString(), "UTF-8"));
-			userId = reqJSON.getInt("userId");
-		} catch (JSONException e) {
-			throw new JSONException("jason error: " + e.getMessage() + " string: " + URLDecoder.decode(request.getQueryString(), "UTF-8"));
 
-		}
 		if (userId > 0) {
 			trips = tripDao.getTrips(userDao.getUser(userId));
 		}
@@ -64,69 +56,54 @@ public class TripController {
 		response.getWriter().write(result );
 	}
 
-	@RequestMapping("/trip/delete.json")
-	public void delete(HttpServletResponse response, HttpServletRequest request)
-			throws JSONException, IOException {
-		JSONObject reqJSON = null;
-		int userId = 0;
-		String tripName = "";
+	private String delete(Integer userId, String tripName) {
 		String result = "";
-		try {
-			reqJSON = new JSONObject(request.getQueryString());
-			userId = reqJSON.getInt("userId");
-			tripName = reqJSON.getString("tripName");
-
-		} catch (JSONException e) {
-			result = "System error";
-			throw new JSONException("jason error: " + request.getQueryString() + " Json: " + reqJSON.toString());
-		}
-
 		if (tripName != "" && userId > 0) {
 			Trip trip = tripDao.getTrip(userDao.getUser(userId), tripName);
-			tripDao.deleteTrip(trip);
-			result = "Trip successfuly removed";
+			if (trip != null) {
+				List<Route> routes = trip.getRoutes();
+				for(int i=0; i < routes.size(); i++) {
+					routeDao.deleteRoute(routes.get(i));
+				}
+				tripDao.deleteTrip(trip);
+				result = "Trip successfuly removed";
+			} else {
+				result = "Trip not found";
+			}
 		} else {
 			result = "Input error";
 		}
-		result = "{\"result\" : \"" + result + "\"}";
-		response.setContentType("application/json");
-		response.getWriter().write(result );
+		return "{\"result\" : \"" + result + "\"}";
 
 	}
 
-	@RequestMapping("/trip/create.json")
-	public void create(HttpServletResponse response, HttpServletRequest request)
+	@RequestMapping("/trip/delete/{userId}/{tripName}")
+	public void delete(@PathVariable Integer userId, @PathVariable String tripName, HttpServletResponse response, HttpServletRequest request)
 			throws JSONException, IOException {
-		JSONObject reqJSON = null;
-		JSONObject tripJSON = null;
+
+		response.setContentType("application/json");
+		response.getWriter().write(delete(userId, URLDecoder.decode(tripName, "UTF-8")));
+
+	}
+
+	private String create(Integer userId, JSONObject tripJSON) {
+		String result = ""; 
 		Trip trip = new Trip();
 		String err = "";
-		String result = "";
-
-		int userId = 0;
-		try {
-			reqJSON = new JSONObject(request.getQueryString());
-			userId = reqJSON.getInt("userId");
-			tripJSON = reqJSON.getJSONObject("trip");
-
-		} catch (JSONException e) {
-			err = "Invalid input";
-			throw new JSONException("jason error: " + request.getQueryString() + " Json: " + reqJSON.toString());
-		}
 
 		if (userId > 0 && tripJSON != null) {
 			trip.setUser(userDao.getUser(userId));
 			try {
 				List<Location> locations = new ArrayList<Location>();
 				JSONArray routeJs = tripJSON.getJSONArray("routes");
-				Location lastLocation = null;
+				Location lastLocation = new Location();
 				trip.setName(tripJSON.getString("name"));
 				trip.setStartTime(Date.valueOf(tripJSON.getString("startTime")));
 				trip.setDefaultTravelMethod(tripJSON.getString("defaultTravelMethod"));
 				trip = tripDao.saveTrip(trip);
 
 				for(int i=0; i< routeJs.length(); i++) {
-					JSONObject routeJ = routeJs.getJSONObject(0);
+					JSONObject routeJ = routeJs.getJSONObject(i);
 					JSONObject locationJ = routeJ.getJSONObject("fromLocation");
 					Route route = new Route();
 					route.setTrip(trip);
@@ -138,6 +115,7 @@ public class TripController {
 					location.setType(locationJ.getString("type"));
 					location.setLat(locationJ.getDouble("lat"));
 					location.setLon(locationJ.getDouble("lon"));
+					location.setNotes(locationJ.getString("notes"));
 					location = locationDao.saveLocation(location);
 
 					route.setFromLocation(location);
@@ -153,6 +131,8 @@ public class TripController {
 					location.setType(locationJ.getString("type"));
 					location.setLat(locationJ.getDouble("lat"));
 					location.setLon(locationJ.getDouble("lon"));
+					location.setNotes(locationJ.getString("notes"));
+
 					location = locationDao.saveLocation(location);
 
 					route.setToLocation(location);
@@ -167,13 +147,60 @@ public class TripController {
 				trip = tripDao.saveTrip(trip);
 			} catch (JSONException e) {
 				err = "Missing parameter";
+			} catch (Exception ex) {
+				err = "others:" + ex.getMessage();
 			}
 		}
 		if (err == "")
 			result = "{\"result\" : \"Trip created\"}";
 		else
 			result = "{\"result\" : \"Error\", \"message\" : \"" + err + "\"}";
+
+		return result;
+	}
+
+	@RequestMapping("/trip/create")
+	public void create(HttpServletResponse response, HttpServletRequest request)
+			throws JSONException, IOException {
+		JSONObject reqJSON = null;
+		String result;
+
+		try {
+			reqJSON = new JSONObject(URLDecoder.decode(request.getQueryString(), "UTF-8"));
+			result = create(reqJSON.getInt("userId"),reqJSON.getJSONObject("trip"));
+
+		} catch (JSONException e) {
+			result = "{\"result\" : \"Invalid input\"}";
+			throw new JSONException("jason error: " + e.getMessage() + " string: " + request.getQueryString());
+		}
+
 		response.setContentType("application/json");
 		response.getWriter().write(result );
 	}
+
+	@RequestMapping("/trip/update")
+	public void update(HttpServletResponse response, HttpServletRequest request)
+			throws JSONException, IOException {
+		JSONObject reqJSON = null;
+		JSONObject tripJSON = null;
+		String result, tripName;
+		Integer userId;
+
+		try {
+			reqJSON = new JSONObject(URLDecoder.decode(request.getQueryString(), "UTF-8"));
+			tripJSON = reqJSON.getJSONObject("trip");
+			userId = reqJSON.getInt("userId");
+			tripName = reqJSON.getString("tripName");
+			result = delete(userId, tripName);
+			if (result != "{\"result\" : \"Input error\"}" && result != "{\"result\" : \"Trip not found\"}")
+				result = create(userId,tripJSON);
+		} catch (JSONException e) {
+			result = "{\"result\" : \"Invalid input\"}";
+			throw new JSONException("jason error: " + e.getMessage() + " string: " + request.getQueryString());
+		}
+
+		response.setContentType("application/json");
+		response.getWriter().write(result );
+	}
+
 }
